api/requirements.txt
aws-lambda-powertools>=2.0.0
boto3>=1.26.0
requests>=2.28.0
PyJWT>=2.6.0
cryptography>=39.0.0
fastjsonschema
pydantic<2.0.0
python-jose>=3.3.0